services/detail_fetcher_service/requirements.txt
fastapi
uvicorn
requests
google-cloud-firestore
google-cloud-storage
google-cloud-bigquery
